feat(batch): add OpenAI/Azure batch support and refactor Vertex batch
